requirements.txt
Flask
Pillow
# pycoral (read official docs), for RPi with python v3.9.2 use
# $ python3 -m pip install https://github.com/google-coral/pycoral/releases/download/v2.0.0/pycoral-2.0.0-cp39-cp39-linux_armv7l.whl
# for MacOS (Armv8) likely need to build from source